lgtm/requirements.txt
Click==7.0
Pillow==6.2.1
requests==2.22.0

